+++ public/r/dashgrid.json
@@ -9,22 +9,22 @@
   ],
   "files": [
     {
-      "path": "registry/blocks/dashgrid/dashgrid.tsx",
+      "path": "registry/blocks/dashgrid/index.tsx",
       "content": "import { createFileRoute } from '@tanstack/react-router';\nimport { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';\nimport { useUsers } from '@/hooks/use-users';\n\nexport const Route = createFileRoute('/dashgrid')({\n  component: RouteComponent\n});\n\nfunction RouteComponent() {\n  const { data: users, isLoading } = useUsers();\n\n  if (isLoading) return <div>Loading...</div>;\n\n  return (\n    <div className=\"container mx-auto p-6\">\n      <h1 className=\"text-2xl font-bold mb-6\">Dashboard</h1>\n      <div className=\"grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4\">\n        {users?.map((user: any) => (\n          <Card key={user.id}>\n            <CardHeader>\n              <CardTitle>{user.name}</CardTitle>\n            </CardHeader>\n            <CardContent>\n              <p>{user.email}</p>\n              <p>{user.company.name}</p>\n            </CardContent>\n          </Card>\n        ))}\n      </div>\n    </div>\n  );\n}\n",
       "type": "registry:page",
-      "target": "~/src/routes/dashgrid.tsx"
+      "target": "~/src/routes/dashgrid/index.tsx"
     },
     {
       "path": "registry/blocks/dashgrid/hooks/use-users.tsx",
       "content": "import { useQuery } from '@tanstack/react-query';\nimport { getUsers } from '@/lib/users';\n\nexport function useUsers() {\n  return useQuery({\n    queryKey: ['users'],\n    queryFn: getUsers\n  });\n}\n",
       "type": "registry:hook",
-      "target": "~/src/hooks/use-users.tsx"
+      "target": "~/src/routes/dashgrid/-hooks/use-users.tsx"
     },
     {
       "path": "registry/blocks/dashgrid/lib/users.tsx",
       "content": "// registry/blocks/dashboard/lib/users.ts\nexport const getUsers = async () => {\n  const response = await fetch('https://jsonplaceholder.typicode.com/users');\n  return response.json();\n};\n",
       "type": "registry:lib",
-      "target": "~/src/lib/users.tsx"
+      "target": "~/src/routes/dashgrid/-lib/users.tsx"
     }
   ]
 }
