+++ src/routes/dashgrid/-hooks/use-users.tsx
@@ -1,6 +1,5 @@
-// registry/blocks/dashboard/hooks/use-users.ts
 import { useQuery } from '@tanstack/react-query';
-import { getUsers } from '@/lib/users';
+import { getUsers } from '../-lib/users';
 
 export function useUsers() {
   return useQuery({
