+++ src/routes/dashgrid/-lib/users.tsx
@@ -1,4 +1,3 @@
-// registry/blocks/dashboard/lib/users.ts
 export const getUsers = async () => {
   const response = await fetch('https://jsonplaceholder.typicode.com/users');
   return response.json();
